pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "youtube-summariser"
version = "0.5.0"
description = "Summarize YouTube videos using AI (OpenAI, Anthropic, or OpenRouter)"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Weijian Zhang" }
]
keywords = [
    "youtube",
    "summarizer",
    "transcript",
    "ai",
    "openai",
    "anthropic",
    "openrouter",
    "claude",
    "gpt",
    "cli"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video",
    "Topic :: Text Processing :: General",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    "openai>=1.60.0",
    "anthropic>=0.40.0",
    "openrouter>=0.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "youtube-transcript-api>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
youtube-summariser = "youtube_summariser.cli:main"

[project.urls]
Homepage = "https://github.com/weijianzhg/youtube-summariser"
Repository = "https://github.com/weijianzhg/youtube-summariser"
Issues = "https://github.com/weijianzhg/youtube-summariser/issues"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/youtube_summariser"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]

